Rename the file, use fs-extra
